added support for single date selection
