+++ android-calendarview-library/src/main/java/com/oyorooms/CalendarCellView.java
@@ -36,9 +36,13 @@ public class CalendarCellView extends AppCompatTextView {
     private static final int[] STATE_PRE_DEFINED_RANGE_MIDDLE = {
             R.attr.state_pre_defined_range_middle
     };
+    private static final int[] STATE_SINGLE_SELECTION = {
+            R.attr.state_single_selection
+    };
 
     private boolean isSelectable = false;
     private boolean isToday = false;
+    private boolean isSingleSelection = false;
 
     private RangeState rangeState = RangeState.NONE;
     private RangeState predefinedRangeState = RangeState.NONE;
@@ -71,6 +75,10 @@ public class CalendarCellView extends AppCompatTextView {
         refreshDrawableState();
     }
 
+    public void setSingleSelection(boolean singleSelection) {
+        isSingleSelection = singleSelection;
+    }
+
     @Override
     protected int[] onCreateDrawableState(int extraSpace) {
         final int[] drawableState = super.onCreateDrawableState(extraSpace + 5);
@@ -85,6 +93,10 @@ public class CalendarCellView extends AppCompatTextView {
             mergeDrawableStates(drawableState, STATE_TODAY);
         }
 
+        if (isSingleSelection){
+            mergeDrawableStates(drawableState, STATE_SINGLE_SELECTION);
+        }
+
         if (rangeState == RangeState.START) {
             mergeDrawableStates(drawableState, STATE_RANGE_START);
         } else if (rangeState == RangeState.MIDDLE) {
@@ -94,7 +106,7 @@ public class CalendarCellView extends AppCompatTextView {
         }
 
         if (predefinedRangeState == RangeState.START){
-            mergeDrawableStates(drawableState, STATE_PREDEFINED_RANGE_START); 
+            mergeDrawableStates(drawableState, STATE_PREDEFINED_RANGE_START);
         } else if (predefinedRangeState == RangeState.MIDDLE) {
             mergeDrawableStates(drawableState, STATE_PRE_DEFINED_RANGE_MIDDLE);
         } else if (predefinedRangeState == RangeState.END) {
