+++ android-calendarview-library/src/main/java/com/oyorooms/DateSelectionListener.java
@@ -6,6 +6,7 @@ import java.util.Date;
  * Created by aneesha.bahukhandi on 16/05/17
  */
 
-public interface RangeSelectionListener {
+public interface DateSelectionListener {
     void onRangeSelected(Date startDate, Date endDate);
+    void onDateSelected(Date date);
 }
