+++ android-calendarview-library/src/main/java/com/oyorooms/DateSelectionMode.java
@@ -1,0 +1,11 @@
+package com.oyorooms;
+
+/**
+ * Created by aneesha.bahukhandi on 17/05/17
+ */
+
+public enum DateSelectionMode {
+
+    SINGLE,
+    RANGE
+}
