--- android-calendarview-library/src/main/res/drawable/calendar_day_of_week_bg.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="@color/colorPrimary"/>
-        </shape>
-    </item>
-    <item android:bottom="6dp">
-        <shape android:shape="rectangle">
-            <solid android:color="@android:color/white"/>
-        </shape>
-    </item>
-</layer-list>
